Rename account file structure
