+++ walk-in-interview/src/main/webapp/account/applicant/edit/index.html
@@ -37,5 +37,5 @@
     </div>
   </body>
 
-  <script type="module" src="./edit-applicant-account.js"></script>
+  <script type="module" src="./script.js"></script>
 </html>
